refactored login/registration forms into modules
